requirements.txt
ffmpeg-python==0.2.0
peft==0.5.0
pycocoevalcap==1.2
tensorboard==2.13.0
transformers==4.46.3
deepspeed==0.15.1
numpy==1.26.4
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
opencv-python-headless==4.10.0.84
pycocotools==2.0.6
mmengine==0.10.6
iopath>=0.1.10
bleach
sentencepiece
scikit-learn
